trying pr

+++ Jenkinsfile
@@ -15,12 +15,18 @@ pipeline {
       agent any
       steps {
         checkout scm
-        sh 'git submodule update --init --recursive'
-        sh 'docker stop $(docker ps -a | grep personal | awk \'{ print $1 }\')'
-        sh 'docker rm $(docker ps -a | grep personal | awk \'{ print $1 }\')'
-        sh 'docker rmi $(docker images | grep personal | awk \'{ print $3 }\')'
-        sh 'docker-compose up --build -d'
-        echo 'successfully deployed'
+        script {
+          if (BRANCH_NAME == "master") {
+            sh 'git submodule update --init --recursive'
+            sh 'docker stop $(docker ps -a | grep personal | awk \'{ print $1 }\')'
+            sh 'docker rm $(docker ps -a | grep personal | awk \'{ print $1 }\')'
+            sh 'docker rmi $(docker images | grep personal | awk \'{ print $3 }\')'
+            sh 'docker-compose up --build -d'
+            echo 'successfully deployed'
+          } else {
+            echo 'Don't have a dev server yet, so just go ahead and push'
+          }
+        }
       }
     }
   }
